update and use size_t, uint and init spheres.h

+++ cube.h
@@ -5,7 +5,7 @@
 #include "vect.h"
 #include <algorithm>
 #include <array>
-#include <cstdint>
+#include <cstddef>
 #include <sys/types.h>
 #include <vector>
 
@@ -13,9 +13,9 @@ class cube : public mesh {
   public:
     cube(double sides) : sides(sides) {
         // use bit & operator to identify the 1.. if a bit is 1 it is value is negative
-        for (int i = 0; i < num_vtx; i++) {
-            for (int j = 0; j < val.size(); j++) {
-                int b = i & (1 << j);
+        for (size_t i = 0; i < num_vtx; i++) {
+            for (size_t j = 0; j < val.size(); j++) {
+                uint b = i & (1 << j);
 
                 if (b == (1 << j)) {
                     val.e[j] = -sides / 2;
@@ -28,24 +28,24 @@ class cube : public mesh {
         }
     };
 
-    std::vector<vect3> surface_interpolation(const uint32_t &subdivision) override {
-        for (int face = 0; face < faces; face++) {
-            int axis = face / 2;  // x = 0; y; 1; z; 2
-            int value = face % 2; // min = 0; max = 1 -> identify the face of
-                                  // which axis we are in the loop
+    std::vector<vect3> surface_interpolation(const uint &subdivision) override {
+        for (size_t face = 0; face < faces; face++) {
+            uint axis = face / 2;  // x = 0; y; 1; z; 2
+            uint value = face % 2; // min = 0; max = 1 -> identify the face of
+                                   // which axis we are in the loop
 
             std::vector<vect3> corners;
 
             // find each corners
-            for (int i = 0; i < vertices.size(); i++) {
+            for (size_t i = 0; i < vertices.size(); i++) {
                 if (((i >> axis) & 1) == value) {
                     corners.push_back(vertices[i]);
                 }
             }
 
             // sort by other axes
-            int axis1 = (axis + 1) % 3;
-            int axis2 = (axis + 2) % 3;
+            uint axis1 = (axis + 1) % 3;
+            uint axis2 = (axis + 2) % 3;
 
             std::sort(corners.begin(), corners.end(),
                       [axis1, axis2](const vect3 a, const vect3 b) -> bool {
@@ -58,8 +58,8 @@ class cube : public mesh {
                       });
 
             // interpolate
-            for (int i = 0; i <= subdivision; i++) {
-                for (int j = 0; j <= subdivision; j++) {
+            for (size_t i = 0; i <= subdivision; i++) {
+                for (size_t j = 0; j <= subdivision; j++) {
                     double u = double(i) / subdivision;
                     double v = double(j) / subdivision;
 
@@ -76,18 +76,18 @@ class cube : public mesh {
         return surface_points;
     }
 
-    std::vector<from_to> grid(uint32_t &subdivision) override {
-        int points_per_face = (subdivision + 1) * (subdivision + 1);
-        int points_per_row = subdivision + 1;
+    std::vector<from_to> grid(uint &subdivision) override {
+        uint points_per_face = (subdivision + 1) * (subdivision + 1);
+        uint points_per_row = subdivision + 1;
         from_to pt;
 
-        for (int face = 0; face < faces; face++) {
-            int face_start = face * points_per_face;
+        for (size_t face = 0; face < faces; face++) {
+            uint face_start = face * points_per_face;
 
-            for (int i = 0; i < points_per_face; i++) {
-                int n = face_start + i;
-                int row = i / points_per_row;
-                int col = i % points_per_row;
+            for (size_t i = 0; i < points_per_face; i++) {
+                uint n = face_start + i;
+                uint row = i / points_per_row;
+                uint col = i % points_per_row;
 
                 if (col < subdivision) {
                     pt[0] = n;
@@ -112,8 +112,8 @@ class cube : public mesh {
     vect3 val;
     double sides = 0.5;
 
-    const uint32_t num_vtx = 8;
-    const uint32_t faces = 6;
+    const uint num_vtx = 8;
+    const uint faces = 6;
 
     std::vector<vect3> vertices;
     std::vector<vect3> surface_points;
